+++ recipes/default.rb
@@ -23,8 +23,7 @@
 # Caddyfile at /etc/Caddyfile
 # pidfile at /var/run/caddy.pid
 
-
-ark 'caddy' do 
+ark 'caddy' do
   url "https://caddyserver.com/download/build?os=linux&arch=amd64&features=#{node['caddy']['features'].join(',')}"
   extension 'tar.gz'
   has_binaries ['./caddy']
@@ -35,7 +34,7 @@ end
 execute 'setcap cap_net_bind_service=+ep caddy' do
   cwd '/usr/local/caddy'
   action :nothing
-  subscribes :run, 'ark[caddy]',:immediately
+  subscribes :run, 'ark[caddy]', :immediately
 end
 
 template '/etc/Caddyfile' do
@@ -55,6 +54,13 @@ if %w(arch gentoo rhel fedora suse).include? node['platform_family']
     mode '0755'
     variables variables
   end
+elsif node['platform'] == 'ubuntu' && node['platform_version'] == '14.04'
+  # Upstart
+  template '/etc/init/caddy.conf' do
+    source 'upstart.erb'
+    mode '0755'
+    variables variables
+  end
 else
   # SysV
   template '/etc/init.d/caddy' do
@@ -66,4 +72,5 @@ end
 
 service 'caddy' do
   action [:enable, :start]
+  supports :status => true, :start => true, :stop => true, :restart => true
 end